start of no drop over card
this is a start to stoping a drag from swapping if a card shouldn't be
swapped.  The problem is that this is way more complicated that
originally imagined.  Moving from one list into new list will require
API changes to block entry when over a specific item.  The entry can be
blocked but there is no current way to then allow it as you move the
drag around in the destination list until you find a card you can drop
over.